+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## \[x.y.z] - Unreleased
 
+### Added
+- New QMI driver for Newport SMC100PP and new actuator UTS100PP in the `actuators.py`. 
+
 ## [0.41.0] - 2023-08-11
 
 ### Added